Tweaking dist files to make @dependendable/state external
