+++ src/index.js
@@ -1,3 +1,2 @@
 export * from "./vdom.js";
-export { h } from "./h.js";
 export { html } from "./html.js";
